various fixes and refactoring * Parameter passing WORKS * DateTime passing don't throw (Need to investigate)
